🔨APP-3926: Restructured the common.schema package into a package with submodules based on functionality use StrEnum & added linter config
